+++ src/ArrayLike.php
@@ -44,7 +44,7 @@ interface ArrayLike extends ArrayAccess, IteratorAggregate, Countable
     /**
      * Reduce the tensor down to a scalar.
      *
-     * @param callable $callback
+     * @param callable $callback function (float $carry, float $value): float
      * @param float $initial
      * @return float
      */